+++ tests/testthat/test-remove.r
@@ -1,0 +1,6 @@
+context("test-remove.r")
+
+test_that("succesfully wraps str_replace_all", {
+  expect_equal(str_remove_all("abababa", "ba"), "a")
+  expect_equal(str_remove("abababa", "ba"), "ababa")
+})